+++ TimeTrackerWeb/Dtos/UserReportDto.cs
@@ -10,9 +10,9 @@ namespace TimeTrackerWeb.Dtos
         public int Id { get; set; }
         public int UserId { get; set; }
         public DateTime Date { get; set; }
-        public float WorkHours { get; set; }
-        public float BreakHours { get; set; }
-        public float TimeDifference { get; set; }
+        public double WorkHours { get; set; }
+        public double BreakHours { get; set; }
+        public double TimeDifference { get; set; }
         public string TimeDifferenceReason { get; set; }
         public int ApprovedByUserId { get; set; }
         public bool ApprovedFlag { get; set; }
